Should not mutate options[:file_name]
The combination of Proc#curry and the `is_a?(Proc)` conditional create a bug that causes the file name to be set once (the first time a file upload is handled by the Rails process) and then never updated again.

Proc#curry's docs state:

> If a sufficient number of arguments are supplied, it passes the supplied arguments to the original proc and returns the result.

Because the Proc's arity is enforced as 1, this means that it will get set to a String after the first run.

This also causes the curious bug of always firing the `file_name` deprecation warning on all but the first request handled by this code.

A better solution is to simply call the proc on demand.